+++ .github/workflows/stale.yml
@@ -13,7 +13,7 @@ jobs:
     env:
       ACTIONS_STEP_DEBUG: true
     steps:
-    - uses: actions/stale@v3.0.10
+    - uses: actions/stale@v3.0.11
       with:
         repo-token: ${{ secrets.GITHUB_TOKEN }}
         stale-issue-message: 'This issue has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs. Thank you for your contributions.'
