--- src/settings.ts
@@ -1,9 +1,0 @@
-/**
- * This is the name of the platform that users will use to register the plugin in the Homebridge config.json
- */
-export const PLATFORM_NAME = 'ExampleHomebridgePlugin';
-
-/**
- * This must match the name of your plugin as defined the package.json
- */
-export const PLUGIN_NAME = 'homebridge-plugin-name';
